# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ElorAdmin API",
        "version": "2.0"
    },
    "paths": {
...
        "/api/modules": {
            "get": {
                "tags": [
                    "Modules"
                ],
                "summary": "Shows modules",
                "operationId": "2e94a3c5aa6628f36f22988ceb189fcc",
                "responses": {
                    "200": {
                        "description": "Shows all modules."
                    },
                    "204": {
                        "description": "There are no modules."
                    },
                    "401": {
                        "description": "You must log in."
                    },
                    "default": {
                        "description": "Error has ocurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Modules"
                ],
                "summary": "Create a module",
                "operationId": "ba4fe3a1c551c9128c95e18e680e2bc7",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Module Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Module code",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hours",
                        "in": "query",
                        "description": "Module duration hours",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cycles",
                        "in": "query",
                        "description": "Cycles",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}